remove auxiliary data uint64 casts & get instance by payment slot

+++ src/aztec/DefiBridgeProxy.sol
@@ -168,7 +168,7 @@ contract DefiBridgeProxy is DSTest {
             outputAssetB,
             totalInputValue,
             interactionNonce,
-            uint64(auxInputData),
+            auxInputData,
             address(0)
         );
